% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{createData}
\alias{createData}
\title{createData}
\usage{
createData(n, sigma_error, example = c(1, 2, 3, 4, 5))
}
\arguments{
\item{n}{number of observations}

\item{sigma_error}{noisiness of the Data Generating Process}

\item{example}{the numeric id of the simulation setting to be used (see report)}
}
\value{
a list containing \code{X}, \code{y} and \code{betas}
}
\description{
creates simulated data to test our solvers on
}
