% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_data_generator.R
\name{createDataSet}
\alias{createDataSet}
\title{createDataSet}
\usage{
createDataSet(p, beta_sparsity, n_train = 20, n_validation = 20,
  n_test = 200, num_datasets = 50, sigma_error)
}
\arguments{
\item{p}{the number of dimensions in the simulated data}

\item{beta_sparsity}{the proportion of true non-zero betas}

\item{n_train}{number of training observations}

\item{n_validation}{number of validation (tuning) observations}

\item{n_test}{number of testing observations}

\item{num_datasets}{number of different datasets to generate}

\item{sigma_error}{noisiness of the Data Generating Process}
}
\value{
a list containing all 50 generated datasets
}
\description{
creates simulated data to test our solvers on
}
