% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossval_EN.R
\name{cv_EN}
\alias{cv_EN}
\title{cv_EN}
\usage{
cv_EN(lambda_max = 10, step_lambda = 0.1, n_folds = 10, y, X,
  one_stderr_rule = TRUE)
}
\arguments{
\item{lambda_max}{maximum lambda to try}

\item{step_lambda}{step size between 0 and \code{lambda_max}}

\item{n_folds}{the number of folds to use in cross-validation}

\item{y}{a nx1 label vector}

\item{X}{the design matrix}

\item{one_stderr_rule}{whether to use the one-standard-error rule when choosing lambda}
}
\value{
a list object containing a data.frame summarizing the error for each pair of lambda/alpha, and the best lambda/alpha
}
\description{
use cross-validation to find the best values of lambda and alpha
}
