% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossval_lasso.R
\name{cv_lasso}
\alias{cv_lasso}
\title{cv_lasso}
\usage{
cv_lasso(lambda_max = 5, step_lambda = 0.1, n_folds = 10, y, X,
  one_stderr_rule = TRUE)
}
\arguments{
\item{lambda_max}{maximum lambda to try}

\item{step_lambda}{step size between 0 and \code{lambda_max}}

\item{n_folds}{the number of folds to use in cross-validation}

\item{y}{a nx1 label vector}

\item{X}{the design matrix}

\item{one_stderr_rule}{whether to use the one-standard-error rule when choosing lambda}
}
\value{
a list object containing a data.frame summarizing the error for each lambda, and the best lambda
}
\description{
use cross-validation to find the best value of lambda
}
