% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossval_EN.R
\name{plot_cv_EN}
\alias{plot_cv_EN}
\title{plot_cv_EN}
\usage{
plot_cv_EN(cv_results)
}
\arguments{
\item{cv_results}{the list returned by \code{cv_EN}}
}
\value{
a plot
}
\description{
plots the result of \code{cv_EN}
}
