% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossval_lasso.R
\name{plot_cv_lasso}
\alias{plot_cv_lasso}
\title{plot_cv.lasso}
\usage{
plot_cv_lasso(cv_results)
}
\arguments{
\item{cv_results}{the list returned by \code{cv_lasso}}
}
\value{
a plot
}
\description{
plots the result of \code{cv_lasso}
}
