% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict}
\alias{predict}
\title{predict}
\usage{
predict(betas, new_data)
}
\arguments{
\item{betas}{the betas returned by a model}

\item{new_data}{new data}
}
\value{
a vector of new predictions
}
\description{
uses the results of a model to predict new data
}
