% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EN_tune.R
\name{tune_EN}
\alias{tune_EN}
\title{tune_EN}
\usage{
tune_EN(lambda_max = 10, step_lambda = 0.1, ytrain, Xtrain, yvalid, Xvalid)
}
\arguments{
\item{lambda_max}{maximum lambda to try}

\item{step_lambda}{step size between 0 and \code{lambda_max}}

\item{ytrain}{a nx1 label vector used for training}

\item{Xtrain}{the design matrix used for training}

\item{yvalid}{a nx1 label vector used for validating}

\item{Xvalid}{the design matrix used for validating}

\item{n_folds}{the number of folds to use in cross-validation}

\item{one_stderr_rule}{whether to use the one-standard-error rule when choosing lambda}
}
\value{
a list object containing a data.frame summarizing the error for each pair of lambda/alpha, and the best lambda/alpha
}
\description{
finds optimal lambda / alpha, given train and validation sets
}
