% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{createData}
\alias{createData}
\title{createData}
\usage{
createData(betas, n, sigma_error)
}
\arguments{
\item{betas}{a vector of beta values}

\item{n}{number of observations}

\item{sigma_error}{noisiness of the Data Generating Process}
}
\value{
a list containing \code{X}, \code{y} and \code{betas}
}
\description{
creates simulated data to test our solvers on
}
